Refactor out `spyOnKafka()`
