+++ plugin-server/tests/helpers/kafka.ts
@@ -1,4 +1,4 @@
-import { Kafka, logLevel } from 'kafkajs'
+import { Consumer, Kafka, logLevel } from 'kafkajs'
 
 import { defaultConfig, overrideWithEnv } from '../../src/config/config'
 import {
@@ -15,9 +15,10 @@ import {
 import { PluginsServerConfig } from '../../src/types'
 import { UUIDT } from '../../src/utils/utils'
 import { KAFKA_EVENTS_DEAD_LETTER_QUEUE } from './../../src/config/kafka-topics'
+import { delayUntilEventIngested } from './clickhouse'
 
-/** Clear the kafka queue */
-export async function resetKafka(extraServerConfig?: Partial<PluginsServerConfig>): Promise<void> {
+/** Clear the Kafka queue and return Kafka object */
+export async function resetKafka(extraServerConfig?: Partial<PluginsServerConfig>): Promise<Kafka> {
     const config = { ...overrideWithEnv(defaultConfig, process.env), ...extraServerConfig }
     const kafka = new Kafka({
         clientId: `plugin-server-test-${new UUIDT()}`,
@@ -37,6 +38,8 @@ export async function resetKafka(extraServerConfig?: Partial<PluginsServerConfig
         KAFKA_PLUGIN_LOG_ENTRIES,
         KAFKA_EVENTS_DEAD_LETTER_QUEUE,
     ])
+
+    return kafka
 }
 
 async function createTopics(kafka: Kafka, topics: string[]) {
@@ -53,4 +56,36 @@ async function createTopics(kafka: Kafka, topics: string[]) {
         })
     }
     await admin.disconnect()
+}
+
+export function spyOnKafka(
+    topic: string,
+    serverConfig?: Partial<PluginsServerConfig>
+): (minLength?: number) => Promise<any[]> {
+    let bufferTopicMessages: any[]
+    let bufferConsumer: Consumer
+
+    beforeAll(async () => {
+        const kafka = await resetKafka(serverConfig)
+        bufferConsumer = kafka.consumer({ groupId: 'e2e-buffer-test' })
+        await bufferConsumer.subscribe({ topic })
+        await bufferConsumer.run({
+            eachMessage: ({ message }) => {
+                const messageValueParsed = JSON.parse(message.value!.toString())
+                bufferTopicMessages.push(messageValueParsed)
+                return Promise.resolve() // Not needed but KafkaJS's typing accepts promises only
+            },
+        })
+    })
+
+    beforeEach(() => {
+        bufferTopicMessages = []
+    })
+
+    afterAll(async () => {
+        await bufferConsumer.stop()
+        await bufferConsumer.disconnect()
+    })
+
+    return async (minLength) => await delayUntilEventIngested(() => bufferTopicMessages, minLength)
 }
